+++ app/src/routes/api/sentry-tunnel/server.test.ts
@@ -1,4 +1,4 @@
-import { forwardToSentry } from './server.testable';
+import { _forwardToSentry } from './+server';
 
 describe('Sentry tunnel', () => {
 	it('fowards to sentry', async () => {
@@ -7,7 +7,7 @@ describe('Sentry tunnel', () => {
 {"event_id":"6e193188f6fd4f9cb647a7e852e7c05b","sent_at":"2023-08-28T14:48:45.788Z","sdk":{"name":"sentry.javascript.svelte","version":"7.46.0"},"dsn":"https://1793e2fe81b04f5aad6e75e0372c9a12@sentry.fabrique.social.gouv.fr/63"}
 {"type":"event"}
 `.trim();
-		await forwardToSentry({
+		await _forwardToSentry({
 			dsn: new URL('https://1793e2fe81b04f5aad6e75e0372c9a12@sentry.fabrique.social.gouv.fr/63'),
 			body,
 			fetch: fakeFetch,
@@ -27,7 +27,7 @@ describe('Sentry tunnel', () => {
 		const fakeFetch = vi.fn();
 		const body = JSON.stringify({ dsn: 'https://unknown.fr' });
 		try {
-			await forwardToSentry({
+			await _forwardToSentry({
 				dsn: new URL('http://lol.com'),
 				body,
 				fetch: fakeFetch,
